Backend/requirements.txt
#pip install -r requirements.txt
flask
flask_restful
beautifulsoup4
flask_cors
google
googlesearch-python